+++ apps/eternal-devotion/src/app/components/offering_selection/selection_layout.module.scss
@@ -1,3 +1,5 @@
+$mobile-breakpoint: 600px;
+
 // Layout
 .grid {
   display: grid;
@@ -21,7 +23,7 @@
     overflow: auto;
   }
 
-  @media (max-width: 600px) {
+  @media (max-width: $mobile-breakpoint) {
     padding: var(--padding--small);
   }
 }
@@ -66,7 +68,7 @@
     opacity: 1;
   }
 
-  @media (max-width: 600px) {
+  @media (max-width: $mobile-breakpoint) {
     &::before,
     &::after {
       width: 100%;
@@ -104,7 +106,7 @@
     display: none;
   }
 
-  @media (max-width: 600px) {
+  @media (max-width: $mobile-breakpoint) {
     scroll-snap-type: none;
     flex-direction: column;
     overflow-x: hidden;
@@ -126,7 +128,7 @@
   gap: 2rem;
   align-self: center;
 
-  @media (max-width: 600px) {
+  @media (max-width: $mobile-breakpoint) {
     display: none;
   }
 }
@@ -135,18 +137,18 @@
   background-color: var(--color--black);
   color: var(--color--white);
   border: none;
-  padding: 0.75rem 1.5rem;
-  font-size: 1.25rem;
+  padding: 0.5rem 1.25rem;
+  font-size: 1.5rem;
   cursor: pointer;
   transition: opacity var(--transition--fast);
   font-family: var(--font--mono);
 
   &:hover {
-    opacity: 0.8;
+    opacity: 0.85;
   }
 
   &:active {
-    opacity: 0.6;
+    opacity: 0.75;
   }
 }
 
@@ -181,7 +183,7 @@
     }
   }
 
-  @media (max-width: 600px) {
+  @media (max-width: $mobile-breakpoint) {
     font-size: 1.5rem;
     padding-top: 1rem;
   }
@@ -193,7 +195,7 @@
   font-style: normal;
   margin-top: 1rem;
 
-  @media (max-width: 600px) {
+  @media (max-width: $mobile-breakpoint) {
     margin-top: 0;
   }
 }
@@ -240,7 +242,7 @@
     }
   }
 
-  @media (max-width: 600px) {
+  @media (max-width: $mobile-breakpoint) {
     width: 100%;
     max-height: 50dvh;
     aspect-ratio: unset;
